+++ fearless_simd_gen/src/arch/sse4_2.rs
@@ -5,6 +5,7 @@ use crate::arch::Arch;
 use crate::types::{ScalarType, VecType};
 use proc_macro2::{Ident, Span, TokenStream};
 use quote::{format_ident, quote};
+use crate::x86_common::{op_suffix, set1_intrinsic, simple_intrinsic};
 
 pub(crate) fn translate_op(op: &str) -> Option<&'static str> {
     Some(match op {
@@ -102,72 +103,4 @@ impl Arch for Sse4_2 {
             }
         }
     }
-}
-
-pub(crate) fn op_suffix(mut ty: ScalarType, bits: usize, sign_aware: bool) -> &'static str {
-    use ScalarType::*;
-    if !sign_aware && ty == Unsigned {
-        ty = Int;
-    }
-    match (ty, bits) {
-        (Float, 32) => "ps",
-        (Float, 64) => "pd",
-        (Float, _) => unimplemented!("{bits} bit floats"),
-        (Int | Mask, 8) => "epi8",
-        (Int | Mask, 16) => "epi16",
-        (Int | Mask, 32) => "epi32",
-        (Int | Mask, 64) => "epi64",
-        (Unsigned, 8) => "epu8",
-        (Unsigned, 16) => "epu16",
-        (Unsigned, 32) => "epu32",
-        (Unsigned, 64) => "epu64",
-        _ => unreachable!(),
-    }
-}
-
-pub(crate) fn set1_intrinsic(ty: ScalarType, bits: usize) -> Ident {
-    use ScalarType::*;
-    let suffix = match (ty, bits) {
-        (Int | Unsigned | Mask, 64) => "epi64x",
-        _ => op_suffix(ty, bits, false),
-    };
-    format_ident!("_mm_set1_{suffix}")
-}
-
-pub(crate) fn simple_intrinsic(name: &str, ty: ScalarType, bits: usize) -> Ident {
-    let suffix = op_suffix(ty, bits, true);
-    format_ident!("_mm_{name}_{suffix}")
-}
-
-pub(crate) fn simple_sign_unaware_intrinsic(name: &str, ty: ScalarType, bits: usize) -> Ident {
-    let suffix = op_suffix(ty, bits, false);
-    format_ident!("_mm_{name}_{suffix}")
-}
-
-pub(crate) fn extend_intrinsic(ty: ScalarType, from_bits: usize, to_bits: usize) -> Ident {
-    let from_suffix = op_suffix(ty, from_bits, true);
-    let to_suffix = op_suffix(ty, to_bits, false);
-    format_ident!("_mm_cvt{from_suffix}_{to_suffix}")
-}
-
-pub(crate) fn cvt_intrinsic(from: VecType, to: VecType) -> Ident {
-    let from_suffix = op_suffix(from.scalar, from.scalar_bits, false);
-    let to_suffix = op_suffix(to.scalar, to.scalar_bits, false);
-    format_ident!("_mm_cvt{from_suffix}_{to_suffix}")
-}
-
-pub(crate) fn pack_intrinsic(from_bits: usize, signed: bool) -> Ident {
-    let unsigned = match signed {
-        true => "",
-        false => "u",
-    };
-    let suffix = op_suffix(ScalarType::Int, from_bits, false);
-    format_ident!("_mm_pack{unsigned}s_{suffix}")
-}
-
-pub(crate) fn unpack_intrinsic(scalar_type: ScalarType, scalar_bits: usize, low: bool) -> Ident {
-    let suffix = op_suffix(scalar_type, scalar_bits, false);
-
-    let low_pref = if low { "lo" } else { "hi" };
-    format_ident!("_mm_unpack{low_pref}_{suffix}")
 }
